scripts/requirements.txt
# for reading the files
pyyaml

# for accessing flight data
xarray
requests
aiohttp
ipfs
ipfsspec
#fsspec!=0.9.0 # 0.9.0 has a bug which leads to incomplete reads via HTTP
#intake[dataframe] # since intake 0.6.1 to_dask() doesn't work anymore without the [dataframe] specification due to a missing msgpack dependency
#intake-xarray>=0.3.2
#pydap

# for decoration
tqdm
coloredlogs
